Create Automated No Pay Report
The original no pay report was a Tableau report, but many of the collectors were confused by it and it ran slowly. 

Using the original SQL written for the report, I pull this data into a data frame, turn it into a collector specific excel file and then email it to that specific collector, Any outliers such as bankruptcies were sent to management to be handled.